backend/pyproject.toml
[project]
name = "multi-agent-rag-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",
    "pyjwt>=2.9.0",

    # LangChain / LangGraph
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",

    # Database
    "supabase>=2.9.0",

    # Document parsing
    "unstructured[pdf,docx,pptx]>=0.16.0",

    # Task queue
    "celery>=5.4.0",
    "redis>=5.2.0",

    # Cache / Rate limiting
    "upstash-redis>=1.3.0",

    # Logging
    "python-json-logger>=2.0.0",
    "logtail-python>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "httpx>=0.28.0",      # for FastAPI TestClient
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
testpaths = ["../tests"]
pythonpath = [".."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
src = ["app", "worker"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
